+++ Changelog.md
@@ -35,6 +35,7 @@ XML specification. See the updated `custom_entities` example!
 
 - [#868]: Allow to have both `$text` and `$value` special fields in one struct. Previously
   any text will be recognized as `$value` field even when `$text` field is also presented.
+- [#868]: Skip text events when deserialize a sequence of items overlapped with text (including CDATA).
 
 ### Misc Changes